+++ sdk/go/vault/tokenauth/_about.go
@@ -1,8 +1,8 @@
 //nolint: lll
-// Package token exports types, functions, subpackages for provisioning token resources.
+// Package tokenauth exports types, functions, subpackages for provisioning tokenauth resources.
 //
 // > This provider is a derived work of the [Terraform Provider](https://github.com/terraform-providers/terraform-provider-vault)
 // > distributed under [MPL 2.0](https://www.mozilla.org/en-US/MPL/2.0/). If you encounter a bug or missing feature,
 // > first check the [`pulumi/pulumi-vault` repo](https://github.com/pulumi/pulumi-vault/issues); however, if that doesn't turn up anything,
 // > please consult the source [`terraform-providers/terraform-provider-vault` repo](https://github.com/terraform-providers/terraform-provider-vault/issues).
-package token
+package tokenauth